+++ lib/doorkeeper/models/access_token_mixin.rb
@@ -223,7 +223,7 @@ module Doorkeeper
           custom_attributes = extract_custom_attributes(token_attributes).presence
           access_token = matching_token_for(
             application, resource_owner, scopes, custom_attributes: custom_attributes, include_expired: false,
-)
+          )
 
           return access_token if access_token&.reusable?
         end
@@ -331,7 +331,7 @@ module Doorkeeper
       def extract_custom_attributes(attributes)
         attributes.with_indifferent_access.slice(
           *Doorkeeper.configuration.custom_access_token_attributes
-)
+        )
       end
     end