+++ lib/doorkeeper/models/concerns/polymorphic_resource_owner.rb
@@ -27,4 +27,3 @@ module Doorkeeper
     end
   end
 end
-